+++ src/components/checkbox/checkbox.types.tsx
@@ -1,0 +1,14 @@
+export const componentPrefix = 'checkbox';
+
+/** Possible checkbox variants */
+export const checkboxVariantArray = ['check', 'indeterminate'] as const;
+
+/** Types derived from arrays */
+export type CheckboxVariant = (typeof checkboxVariantArray)[number];
+
+/** Base props for the checkbox component */
+export interface CheckboxBaseProps {
+  name?: string;
+  label?: string;
+  variant?: CheckboxVariant;
+}