+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Todas as mudanças notáveis deste projeto serão documentadas neste arquivo.
 
 Este projeto segue o [Keep a Changelog](https://keepachangelog.com/pt-BR/1.0.0/) e [Semantic Versioning](https://semver.org/lang/pt-BR/).
 
+## [2.9.13] - 2026-01-21
+
+### 🚀 Adicionado
+
+- **Checkbox**: Criado componente
+- **Radio**: Criado componente
+
 ## [2.9.12] - 2026-01-21
 
 ### 🚀 Adicionado